setup-environment-internal: do not fait when EULA is not used
The logic we used to detect whether a EULA is needed or not relies on:

find ../sources -print | grep "conf/eula/$MACHINE"

However, when there is no EULA for a specific MACHINE, grep will return
an error code, and if setup-environment script is called with 'set -e', then it
will fail/abort, which is unexpected. We need this specific command to always
return 0, whether the EULA exists or not, this is what this commit is about.

Change-Id: I2520db42d6232ff12eee841b1ef1bf71fe210443
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>