setup-environment-internal: do not unset MACHINE, SDKMACHINE and DISTRO
in aca459cd394112e10f3a5107dfcb529c9ca8f05e, we unset a lot of variables which
were leaked in the environment, however MACHINE, SDKMACHINE and DISTRO are
possibly 'input' variables for the setup-environment script, they are not local
variables that we use temporarily so they should be preserved in case the caller
had set them up.

For example, it is valid to do:

$ export MACHINE=xxx
$ export DISTRO=zzz
$ source setup-environment

In such a case the user wouldn't expect that we unset them.

Change-Id: I11083cfd73c43d861dc44138032a80862bd44f50
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>