setup-environment-internal: use --menu instead of --radiolist for whiptail
Using radiolist is confusing and error prone. The user has to press the space
bar to select an item, which might not be obvious. And if the user presses
'Enter' first, then the value is not set.

There is no obvious reason to use radiolist, since what we have really is a
menu. The default item will be the first one, but at least we do not end up in a
situation where no value is set.

Change-Id: If00c2040fb94573c732978e9d7c746e16253ea2a
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>