README.md: remove leading '$'
With the '$' sign on each command line , it makes it inconvenient to copy/paste
from the web into a terminal. Since commands are already using the proper markup
tags, they are properly highlighted on the web page, so removing the '$' sign
doesn't make a bad visual impact and let users copy/paste more conveniently.

Change-Id: Id48af8f3b91e9464cd1d3e29b343b97340f0f7b1
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>