Update release notes schema to per-update sequence.
To prevent us from re-releasing old release notes, update the schema
and release notes file so they are a sequence of version-numbered
updates, which we can append onto. This lets us see the history of
updates over time, and also has the benefit of letting us reconstruct
release notes for older updates, etc.

Delete the "empty" notes and replace them with a template,
release-notes.template.yml, that can be copy/pasted into
the top of release-notes.yml and edited appropriately.

Change-Id: Ic81030dc6b9f4464c0578560034a7d0f23892818
Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>